Course screen  ui edites

+++ LTA_app/Client/src/com/fym/lta/ui/CourseScreen.java
@@ -97,11 +97,16 @@ private CourseBao courseobjBao = new BaoFactory().createCourseBao();
             }
         });
 
-        jLabel1.setFont(new java.awt.Font("Tahoma", 1, 18)); // NOI18N
+        search_textField.setFont(new java.awt.Font("Tahoma", 0, 14)); // NOI18N
+        search_textField.setBorder(javax.swing.BorderFactory.createLineBorder(new java.awt.Color(0, 0, 0)));
+
+        jLabel1.setFont(new java.awt.Font("Times New Roman", 1, 36)); // NOI18N
         jLabel1.setText("Course Screen ");
 
         jLabel2.setFont(new java.awt.Font("Tahoma", 1, 18)); // NOI18N
-        jLabel2.setText("Course Name");
+        jLabel2.setText("Search");
+
+        jPanel1.setBorder(javax.swing.BorderFactory.createTitledBorder(null, "Courses", 0, 0, new java.awt.Font("Tekton Pro Cond", 1, 36))); // NOI18N
 
         updateCourseBtn.setFont(new java.awt.Font("Tahoma", 1, 18)); // NOI18N
         updateCourseBtn.setText("Update");
@@ -120,13 +125,14 @@ private CourseBao courseobjBao = new BaoFactory().createCourseBao();
         });
 
         insertCourseBtn.setFont(new java.awt.Font("Tahoma", 1, 18)); // NOI18N
-        insertCourseBtn.setText("Insert New Course");
+        insertCourseBtn.setText("Insert New ");
         insertCourseBtn.addMouseListener(new java.awt.event.MouseAdapter() {
             public void mouseClicked(java.awt.event.MouseEvent evt) {
                 insertCourseBtnMouseClicked(evt);
             }
         });
 
+        courseTable.setFont(new java.awt.Font("Tekton Pro Cond", 1, 18)); // NOI18N
         courseTable.setModel(new javax.swing.table.DefaultTableModel(
             new Object [][] {
 
@@ -135,11 +141,12 @@ private CourseBao courseobjBao = new BaoFactory().createCourseBao();
                 "Name", "Code", "Lecture Time ", "Department", "preferred location type"
             }
         ));
+        courseTable.setRowHeight(30);
         courseTable.addHierarchyBoundsListener(new java.awt.event.HierarchyBoundsListener() {
-            public void ancestorMoved(java.awt.event.HierarchyEvent evt) {
-            }
             public void ancestorResized(java.awt.event.HierarchyEvent evt) {
                 courseTableAncestorResized(evt);
+            }
+            public void ancestorMoved(java.awt.event.HierarchyEvent evt) {
             }
         });
         jScrollPane1.setViewportView(courseTable);
@@ -153,17 +160,16 @@ private CourseBao courseobjBao = new BaoFactory().createCourseBao();
         jPanel1.setLayout(jPanel1Layout);
         jPanel1Layout.setHorizontalGroup(
             jPanel1Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGroup(javax.swing.GroupLayout.Alignment.TRAILING, jPanel1Layout.createSequentialGroup()
-                .addGap(52, 52, 52)
-                .addComponent(updateCourseBtn, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
-                .addGap(39, 39, 39)
-                .addComponent(deleteCourseBtn)
-                .addGap(44, 44, 44)
-                .addComponent(insertCourseBtn)
-                .addGap(103, 103, 103))
             .addGroup(jPanel1Layout.createSequentialGroup()
                 .addContainerGap()
-                .addComponent(jScrollPane1)
+                .addGroup(jPanel1Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                    .addGroup(javax.swing.GroupLayout.Alignment.TRAILING, jPanel1Layout.createSequentialGroup()
+                        .addComponent(updateCourseBtn, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+                        .addGap(103, 103, 103)
+                        .addComponent(deleteCourseBtn)
+                        .addGap(112, 112, 112)
+                        .addComponent(insertCourseBtn))
+                    .addComponent(jScrollPane1))
                 .addContainerGap())
         );
         jPanel1Layout.setVerticalGroup(
@@ -175,7 +181,7 @@ private CourseBao courseobjBao = new BaoFactory().createCourseBao();
                     .addComponent(deleteCourseBtn)
                     .addComponent(insertCourseBtn))
                 .addGap(18, 18, 18)
-                .addComponent(jScrollPane1, javax.swing.GroupLayout.DEFAULT_SIZE, 818, Short.MAX_VALUE)
+                .addComponent(jScrollPane1, javax.swing.GroupLayout.DEFAULT_SIZE, 802, Short.MAX_VALUE)
                 .addContainerGap())
         );
 
@@ -203,15 +209,22 @@ private CourseBao courseobjBao = new BaoFactory().createCourseBao();
             }
         });
 
+        code_textField.setFont(new java.awt.Font("Tahoma", 0, 14)); // NOI18N
+
+        name_textField.setFont(new java.awt.Font("Tahoma", 0, 14)); // NOI18N
+
         prefered_ComboBox.setFont(new java.awt.Font("Tahoma", 1, 18)); // NOI18N
 
         Departmentlapel1.setFont(new java.awt.Font("Tahoma", 1, 18)); // NOI18N
         Departmentlapel1.setText("pref_Space");
 
+        hours_perWeak.setFont(new java.awt.Font("Tahoma", 0, 14)); // NOI18N
+
         namelapel1.setFont(new java.awt.Font("Tahoma", 1, 18)); // NOI18N
         namelapel1.setText("Hours per Week");
 
         descrption.setColumns(20);
+        descrption.setFont(new java.awt.Font("Tahoma", 0, 14)); // NOI18N
         descrption.setRows(5);
         jScrollPane2.setViewportView(descrption);
 
@@ -268,11 +281,14 @@ private CourseBao courseobjBao = new BaoFactory().createCourseBao();
                 .addGroup(insertPanelLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
                     .addComponent(codelapel, javax.swing.GroupLayout.PREFERRED_SIZE, 34, javax.swing.GroupLayout.PREFERRED_SIZE)
                     .addComponent(code_textField, javax.swing.GroupLayout.PREFERRED_SIZE, 35, javax.swing.GroupLayout.PREFERRED_SIZE))
-                .addGap(26, 26, 26)
-                .addGroup(insertPanelLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
-                    .addComponent(name_textField, javax.swing.GroupLayout.DEFAULT_SIZE, 32, Short.MAX_VALUE)
-                    .addComponent(namelapel, javax.swing.GroupLayout.PREFERRED_SIZE, 33, javax.swing.GroupLayout.PREFERRED_SIZE))
-                .addGap(33, 33, 33)
+                .addGroup(insertPanelLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                    .addGroup(insertPanelLayout.createSequentialGroup()
+                        .addGap(26, 26, 26)
+                        .addComponent(namelapel, javax.swing.GroupLayout.PREFERRED_SIZE, 33, javax.swing.GroupLayout.PREFERRED_SIZE))
+                    .addGroup(insertPanelLayout.createSequentialGroup()
+                        .addGap(18, 18, 18)
+                        .addComponent(name_textField, javax.swing.GroupLayout.PREFERRED_SIZE, 33, javax.swing.GroupLayout.PREFERRED_SIZE)))
+                .addGap(28, 28, 28)
                 .addGroup(insertPanelLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
                     .addComponent(hours_perWeak, javax.swing.GroupLayout.PREFERRED_SIZE, 32, javax.swing.GroupLayout.PREFERRED_SIZE)
                     .addComponent(namelapel1, javax.swing.GroupLayout.PREFERRED_SIZE, 33, javax.swing.GroupLayout.PREFERRED_SIZE))
@@ -293,7 +309,7 @@ private CourseBao courseobjBao = new BaoFactory().createCourseBao();
                         .addComponent(Departmentlapel2)))
                 .addGap(57, 57, 57)
                 .addComponent(btnSave, javax.swing.GroupLayout.PREFERRED_SIZE, 50, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addGap(37, 37, 37))
+                .addContainerGap(javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))
         );
 
         javax.swing.GroupLayout layout = new javax.swing.GroupLayout(this);
@@ -301,40 +317,43 @@ private CourseBao courseobjBao = new BaoFactory().createCourseBao();
         layout.setHorizontalGroup(
             layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
             .addGroup(layout.createSequentialGroup()
-                .addContainerGap(22, Short.MAX_VALUE)
-                .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                .addContainerGap(javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+                .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING, false)
+                    .addGroup(layout.createSequentialGroup()
+                        .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                            .addComponent(jPanel1, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
+                            .addGroup(layout.createSequentialGroup()
+                                .addComponent(search_textField, javax.swing.GroupLayout.PREFERRED_SIZE, 229, javax.swing.GroupLayout.PREFERRED_SIZE)
+                                .addGap(30, 30, 30)
+                                .addComponent(btnSearch)))
+                        .addGap(18, 18, 18)
+                        .addComponent(insertPanel, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
+                        .addGap(69, 69, 69))
                     .addGroup(layout.createSequentialGroup()
                         .addComponent(jLabel2, javax.swing.GroupLayout.PREFERRED_SIZE, 163, javax.swing.GroupLayout.PREFERRED_SIZE)
-                        .addGap(70, 70, 70)
-                        .addComponent(jLabel1, javax.swing.GroupLayout.PREFERRED_SIZE, 211, javax.swing.GroupLayout.PREFERRED_SIZE))
-                    .addComponent(jPanel1, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                    .addGroup(layout.createSequentialGroup()
-                        .addComponent(search_textField, javax.swing.GroupLayout.PREFERRED_SIZE, 229, javax.swing.GroupLayout.PREFERRED_SIZE)
-                        .addGap(224, 224, 224)
-                        .addComponent(btnSearch)))
-                .addGap(18, 18, 18)
-                .addComponent(insertPanel, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addGap(69, 69, 69))
+                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+                        .addComponent(jLabel1, javax.swing.GroupLayout.PREFERRED_SIZE, 319, javax.swing.GroupLayout.PREFERRED_SIZE)
+                        .addGap(468, 468, 468))))
         );
         layout.setVerticalGroup(
             layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
             .addGroup(layout.createSequentialGroup()
                 .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
                     .addGroup(layout.createSequentialGroup()
-                        .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-                            .addGroup(layout.createSequentialGroup()
-                                .addGap(29, 29, 29)
-                                .addComponent(jLabel2, javax.swing.GroupLayout.PREFERRED_SIZE, 22, javax.swing.GroupLayout.PREFERRED_SIZE))
-                            .addComponent(jLabel1, javax.swing.GroupLayout.PREFERRED_SIZE, 39, javax.swing.GroupLayout.PREFERRED_SIZE))
-                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                        .addGap(29, 29, 29)
+                        .addComponent(jLabel2, javax.swing.GroupLayout.PREFERRED_SIZE, 22, javax.swing.GroupLayout.PREFERRED_SIZE))
+                    .addComponent(jLabel1, javax.swing.GroupLayout.PREFERRED_SIZE, 39, javax.swing.GroupLayout.PREFERRED_SIZE))
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.UNRELATED)
+                .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                    .addGroup(layout.createSequentialGroup()
                         .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
-                            .addComponent(btnSearch, javax.swing.GroupLayout.PREFERRED_SIZE, 33, javax.swing.GroupLayout.PREFERRED_SIZE)
-                            .addComponent(search_textField, javax.swing.GroupLayout.PREFERRED_SIZE, 33, javax.swing.GroupLayout.PREFERRED_SIZE))
-                        .addGap(30, 30, 30)
-                        .addComponent(jPanel1, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))
+                            .addComponent(search_textField, javax.swing.GroupLayout.PREFERRED_SIZE, 33, javax.swing.GroupLayout.PREFERRED_SIZE)
+                            .addComponent(btnSearch, javax.swing.GroupLayout.PREFERRED_SIZE, 48, javax.swing.GroupLayout.PREFERRED_SIZE))
+                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED, 20, Short.MAX_VALUE)
+                        .addComponent(jPanel1, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE))
                     .addGroup(layout.createSequentialGroup()
                         .addComponent(insertPanel, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                        .addGap(0, 332, Short.MAX_VALUE)))
+                        .addGap(0, 0, Short.MAX_VALUE)))
                 .addContainerGap())
         );
     }//GEN-END:initComponents
@@ -350,9 +369,6 @@ private CourseBao courseobjBao = new BaoFactory().createCourseBao();
 
     private void btnSearchMouseClicked(java.awt.event.MouseEvent evt) {//GEN-FIRST:event_btnSearchMouseClicked
     insertPanel.setVisible(false);
-        
-        code_textField.setEnabled(true);
-        
     if (search_textField != null){
         
         CourseDto course = new CourseDto();
@@ -370,13 +386,12 @@ private CourseBao courseobjBao = new BaoFactory().createCourseBao();
     private void insertCourseBtnMouseClicked(java.awt.event.MouseEvent evt) {//GEN-FIRST:event_insertCourseBtnMouseClicked
         insertPanel.setVisible(true);
         updateFlag = false ;
-        // clear data 
+        code_textField.setEnabled(true);
  
         
     }//GEN-LAST:event_insertCourseBtnMouseClicked
 
     private void btnSaveMouseClicked(java.awt.event.MouseEvent evt) {//GEN-FIRST:event_btnSaveMouseClicked
-     String preferedSelectedvType = prefered_ComboBox.getSelectedItem().toString();
     
      DepartmentDto department = new DepartmentDto();
      department.setDepartment_id(allDepartments.get(Department_ComboBox.getSelectedIndex()).getDepartment_id());
@@ -385,7 +400,11 @@ private CourseBao courseobjBao = new BaoFactory().createCourseBao();
     CourseDto course = new CourseDto();
     course.setName(name_textField.getText());
     course.setCode(code_textField.getText());
-    course.setHoursperWeak(Integer.parseInt(hours_perWeak.getText()));
+        try {
+            course.setHoursperWeak(Integer.parseInt(hours_perWeak.getText()));
+        } catch (NumberFormatException nfe) {
+            JOptionPane.showMessageDialog(this, "Hours should intered as a numeric value");
+        }
     course.setDescription(descrption.getText());
 
     try {
@@ -434,18 +453,24 @@ private CourseBao courseobjBao = new BaoFactory().createCourseBao();
     private void deleteCourseBtnActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_deleteCourseBtnActionPerformed
          code_textField.setEnabled(true);
     
-        // get the id value of the selected row 
-        String selectedCourse=  courseTable.getValueAt(courseTable.getSelectedRow(), 1).toString() ;
-       CourseDto selectedCourse_Delete = new CourseDto(); // this course i want to delete
-       selectedCourse_Delete.setCode(selectedCourse);
-       
-       if (courseobjBao.deleteCourse(selectedCourse_Delete)) {
-            JOptionPane.showMessageDialog(this , "deleted successfully");
-            courseTableReset(courseobjBao.listCourses() );
+        if(courseTable.getSelectedRow() >=0)
+        {
+            // get the id value of the selected row 
+            String selectedCourse=  courseTable.getValueAt(courseTable.getSelectedRow(), 1).toString() ;
+            CourseDto selectedCourse_Delete = new CourseDto(); // this course i want to delete
+            selectedCourse_Delete.setCode(selectedCourse);
+            int msg=  JOptionPane.showConfirmDialog(this, "are you sure you need to delete ");
+                         if (msg == JOptionPane.OK_OPTION)
+                         {
+            if (courseobjBao.deleteCourse(selectedCourse_Delete)) {
+                JOptionPane.showMessageDialog(this , "deleted successfully");
+                courseTableReset(courseobjBao.listCourses() );
+            }
+            
+            else
+               JOptionPane.showMessageDialog(this, "Can not delete may be deleted using another Course "); }   
         }
-       
-    else
-           JOptionPane.showMessageDialog(this, "Can not delete may be deleted using another Course ");
+        else JOptionPane.showMessageDialog(this, "Select a Course to Delete");
       
     }//GEN-LAST:event_deleteCourseBtnActionPerformed
 
@@ -501,7 +526,7 @@ private CourseBao courseobjBao = new BaoFactory().createCourseBao();
     
     public void courseTableReset(List<CourseDto> courses) {
        
-        Object[][] coursesArr = new Object[courses.size()][4];
+        Object[][] coursesArr = new Object[courses.size()][8];
 
         for (int i = 0; i < courses.size(); i++) {
 
@@ -509,13 +534,18 @@ private CourseBao courseobjBao = new BaoFactory().createCourseBao();
             coursesArr[i][1] = courses.get(i).getCode();
             coursesArr[i][2] = courses.get(i).getHoursperWeak();
             coursesArr[i][3] = courses.get(i).getDescription();
+            coursesArr[i][4] = courses.get(i).getInsertedBy();
+            coursesArr[i][5] = courses.get(i).getInertion_Date();
+            coursesArr[i][6] = courses.get(i).getUpdatedBy();
+            coursesArr[i][7] = courses.get(i).getUpdate_Date();
 
            
 
         }
       courseTable.setModel(new javax.swing.table.DefaultTableModel(coursesArr, new String[] {
                                                                      "Name", "code" ,
-                                                                    "Hours per Weak ","Description"
+                                                                    "Hours per Weak ","Description","Inserted By",
+ "  Insertion Date", "Updated By", "Update Date"
             }
                                                                    ));
     }
